refactor: 빌드에러 수정 및 폴더구조 수정
