+++ src/pages/CreateBoardGameTip/hooks/useCreateBoardGameTip.ts
@@ -1,5 +1,5 @@
-import { usePostBoardGameTipApi } from '@/apis/postBoardGameTip';
-import { CreateTipRequest } from '@/apis/postBoardGameTip';
+import { usePostBoardGameTipApi } from '@/apis/boardGameTip';
+import { CreateTipRequest } from '@/apis/boardGameTip';
 
 export const useCreateBoardGameTip = () => {
   const api = usePostBoardGameTipApi();
